ModalAlert, Button: add dataTestId
